tests/wpt/web-platform-tests/css/css-fonts/parsing/font-language-override-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-language-override with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#propdef-font-language-override">
<meta name="assert" content="font-language-override supports the full grammar 'normal | <string>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-language-override', 'normal');
test_valid_value('font-language-override', '"ksw"');
</script>
</body>
</html>
